data_ingestion/periodic/requirements.txt
python-dotenv
vnstock3
psycopg2